Organize and purge basics from attributes form
