--- frontend/src/components/attribute_form/background_form/background_form.jsx
@@ -1,9 +1,0 @@
-import React, { useEffect, useState } from 'react';
-
-const BackgroundForm = props => {
-  return <>
-  bg form
-  </>
-};
-
-export default BackgroundForm;
