--- frontend/src/components/attribute_form/background_form/background_form_container.js
@@ -1,15 +1,0 @@
-import { connect } from 'react-redux';
-import BackgroundForm from './background_form.jsx';
-import { closeModal } from '../../../actions/modal_actions';
-import { editUser } from '../../../actions/user_actions';
-
-const mSTP = state => ({
-  user: state.entities.users.user
-});
-
-const mDTP = dispatch => ({
-  closeModal: () => dispatch(closeModal()),
-  editUser: user => dispatch(editUser(user))
-});
-
-export default connect(mSTP, mDTP)(BackgroundForm);
